@@ -11,6 +11,9 @@
   },
   "dependencies": {
     "@next/mdx": "^14.2.5",
+    "@payloadcms/db-postgres": "^3.60.0",
+    "@payloadcms/next": "^3.60.0",
+    "@payloadcms/richtext-lexical": "^3.60.0",
     "@radix-ui/react-accordion": "^1.1.2",
     "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-slot": "^1.0.2",
@@ -21,14 +24,17 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "framer-motion": "^11.18.2",
+    "graphql": "^16.11.0",
     "leva": "^0.10.0",
     "lucide-react": "^0.400.0",
-    "next": "14.2.5",
+    "next": "^15.5.6",
     "next-mdx-remote": "^5.0.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "payload": "^3.60.0",
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0",
     "react-intersection-observer": "^9.16.0",
     "resend": "^3.4.0",
+    "sharp": "^0.34.4",
     "tailwind-merge": "^2.4.0",
     "tailwindcss-animate": "^1.0.7",
     "three": "^0.165.0"
@@ -36,8 +42,8 @@
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.13",
     "@types/node": "^20.19.17",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/react": "^19.2.2",
+    "@types/react-dom": "^19.2.2",
     "autoprefixer": "^10.4.19",
     "eslint": "^8",
     "eslint-config-next": "14.2.5",
