@@ -83,11 +83,12 @@ const blogPosts = [
 ]
 
 type Props = {
-  params: { slug: string }
+  params: Promise<{ slug: string }>
 }
 
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
-  const post = blogPosts.find(p => p.slug === params.slug)
+  const { slug } = await params
+  const post = blogPosts.find(p => p.slug === slug)
 
   if (!post) {
     return {
@@ -108,8 +109,9 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
   }
 }
 
-export default function BlogPost({ params }: Props) {
-  const post = blogPosts.find(p => p.slug === params.slug)
+export default async function BlogPost({ params }: Props) {
+  const { slug } = await params
+  const post = blogPosts.find(p => p.slug === slug)
 
   if (!post) {
     notFound()
