@@ -3,51 +3,52 @@ import { generateMetadata as generateSEO } from '@/lib/seo'
 import { CTABand } from '@/components/sections/CTABand'
 
 export const metadata: Metadata = generateSEO({
-  title: 'Affair Recovery Testimonials | San Diego Couples Counseling',
-  description: 'Real client testimonials about affair recovery therapy in San Diego—how couples rebuilt trust, calmed emotions, and reconnected.',
+  title: 'Client Success Stories | Couples Therapy Reviews',
+  description: 'Read how couples have successfully rebuilt after affairs. Real stories of healing and trust restoration through therapy.',
+  keywords: ['couples therapy reviews san diego', 'marriage counseling testimonials'],
   url: '/testimonials',
 })
 
 const testimonials = [
   {
-    quote: "Jordan gave us the roadmap we desperately needed. We're not just surviving—we're closer than ever.",
-    author: "M.S.",
+    quote: "We are so happy we connected with Jordan. We were reluctant to try therapy at first but after meeting him, we felt confident and comfortable.",
+    author: "Nichole A.",
     role: "Couples Client"
   },
   {
-    quote: "He was the only one who could get through to my husband. Firm, but kind. Exactly what we needed.",
-    author: "L.B.",
+    quote: "I didn't really notice how [anxiety] was impacting my life until my wife told me she was worried about our marriage and my overall well being. She suggested I tried therapy, but I was hesitant... Jordan helped me work through everything and our marriage is stronger than ever.",
+    author: "Peter C.",
     role: "Couples Client"
   },
   {
-    quote: "I thought trust was gone for good. Jordan showed us how to rebuild it, step by step.",
-    author: "J.P.",
+    quote: "Jordan has helped so much with dealing with my husband. I would highly recommend.",
+    author: "Darlene A.",
     role: "Couples Client"
   },
   {
-    quote: "Jordan doesn't just listen—he teaches you real tools.",
-    author: "C.W.",
-    role: "Couples Client"
+    quote: "He's helped me immensely with ways to manage and improve my anxiety with various coping mechanisms. He helps me spread my thoughts out and rationalize and realize. He's easy to talk to as well as patient and kind. Highly recommend.",
+    author: "Joseph F.",
+    role: "Individual Client"
   },
   {
-    quote: "For the first time, I felt like a therapist understood both of us.",
-    author: "A.K.",
-    role: "Couples Client"
+    quote: "After a traumatic accident, I started therapy with Jordan. Jordan was easy to talk to, and helped me process what was going on, and help me cope with the trauma.",
+    author: "Kevin M.",
+    role: "Individual Client"
   },
   {
-    quote: "He made me feel safe and held my partner accountable without shaming him.",
-    author: "R.T.",
-    role: "Couples Client"
+    quote: "Jordan has been so helpful to me! He has been able to provide me honest and constructive feedback in a way that has truly made a positive impact on me. Jordan has a fantastic ability to really listen to what you are saying.",
+    author: "V.S.",
+    role: "Individual Client"
   },
   {
-    quote: "Jordan gave us structure, hope, and confidence we could make it through.",
-    author: "E.B.",
-    role: "Couples Client"
+    quote: "Changed my life! He has made a huge impact. I struggle with severe trauma and depression, he has helped me deal through hard moments. He gets me and not only does he give me guidance, he meets me where I'm at and helps me get back up.",
+    author: "Ale V.",
+    role: "Individual Client"
   },
   {
-    quote: "We walked in doubting everything. We walked out with hope—and over time, trust.",
-    author: "D.L.",
-    role: "Couples Client"
+    quote: "Jordan is incredibly welcoming, warm, intelligent, intuitive and compassionate. I have witnessed his kind hearted approach. He is great in making you feel comfortable and safe.",
+    author: "Alex G.",
+    role: "Colleague Recommendation"
   },
 ]
 
